src/main/java/ejercicios/AnnosBisiestos.java

package ejercicios;

import static ejercicios.SetDeTenis.evaluar;
import java.util.Scanner;

/**
 *
 * @author danielsanchez
 */
public class AnnosBisiestos {
    public static String evaluar(int anno){
        int annobis = anno%4;
        
        if (annobis==0){
            String res= (anno+" es bisiesto");
            return res;
        }else{
            String res2= (anno+" no es bisiesto");
            return res2;
        }
    }
    
    public static void main(String[] args) {
        Scanner lector = new Scanner(System.in);
        System.out.print("Año:");
        int anno = lector.nextInt();
        
        String respuesta = evaluar(anno);
        System.out.println(respuesta);
    }
            
}
